+++ src/components/tooltips/Tooltip.test.tsx
@@ -22,6 +22,7 @@ import { compositeA, compositeB, compositeC } from '../../__mocks__/models/compo
 describe('Test Tooltips', () => {
   let renderTime: Date;
   let props: TooltipProps;
+  let timeZone = 'utc';
   beforeAll(() => {
     renderTime = new Date('01 October 2022 10:28 UTC');
     props = {
@@ -47,7 +48,7 @@ describe('Test Tooltips', () => {
 
   describe('Tooltip Sorting', () => {
     it('returns unsorted metrics', () => {
-      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: applied[0],
         valueEnabled: true,
@@ -72,7 +73,7 @@ describe('Test Tooltips', () => {
       expect(rows[5].innerHTML).toContain('C-series');
     });
     it('returns primary sorted metrics: case sensitive ascending', () => {
-      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: applied[0],
         valueEnabled: true,
@@ -98,7 +99,7 @@ describe('Test Tooltips', () => {
     });
 
     it('returns primary sorted metrics: case sensitive descending', () => {
-      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: applied[0],
         valueEnabled: true,
@@ -125,7 +126,7 @@ describe('Test Tooltips', () => {
 
     it('returns primary sorted metrics: numerical ascending', () => {
       let internalData = [numericalModelA, numericalModelB, numericalModelC];
-      internalData = ApplyComposites([compositeC], internalData, (val) => val, false);
+      internalData = ApplyComposites([compositeC], internalData, (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: internalData[0],
         valueEnabled: true,
@@ -152,7 +153,7 @@ describe('Test Tooltips', () => {
 
     it('returns primary sorted metrics: numerical descending', () => {
       let internalData = [numericalModelA, numericalModelB, numericalModelC];
-      internalData = ApplyComposites([compositeC], internalData, (val) => val, false);
+      internalData = ApplyComposites([compositeC], internalData, (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: internalData[0],
         valueEnabled: true,
@@ -178,7 +179,7 @@ describe('Test Tooltips', () => {
     });
 
     it('returns primary sorted metrics: case insensitive ascending', () => {
-      const applied = ApplyComposites([compositeA], [casedModelA, casedModelB, casedModelC], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [casedModelA, casedModelB, casedModelC], (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: applied[0],
         valueEnabled: true,
@@ -204,7 +205,7 @@ describe('Test Tooltips', () => {
     });
 
     it('returns primary sorted metrics: case insensitive descending', () => {
-      const applied = ApplyComposites([compositeA], [casedModelA, casedModelB, casedModelC], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [casedModelA, casedModelB, casedModelC], (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: applied[0],
         valueEnabled: true,
@@ -232,7 +233,7 @@ describe('Test Tooltips', () => {
 
   describe('Tooltip Generation', () => {
     it('returns tooltip for single metric', () => {
-      const applied = ApplyComposites([compositeA], [modelA], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA], (val) => val, false, timeZone);
       const props: TooltipProps = {
         data: applied[0],
         valueEnabled: true,
@@ -252,19 +253,19 @@ describe('Test Tooltips', () => {
       expect(screen.getByRole('table')).toMatchSnapshot();
     });
     it('returns tooltip for composite metric with one metric', () => {
-      const applied = ApplyComposites([compositeA], [modelA], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA], (val) => val, false, timeZone);
       props.data = applied[0];
       render(<Tooltip {...props} />);
       expect(screen.getByRole('table')).toMatchSnapshot();
     });
     it('returns tooltip for composite metric with two metrics', () => {
-      const applied = ApplyComposites([compositeA], [modelA, modelB], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA, modelB], (val) => val, false, timeZone);
       props.data = applied[0];
       render(<Tooltip {...props} />);
       expect(screen.getByRole('table')).toMatchSnapshot();
     });
     it('returns tooltip for composite metric with three metrics', () => {
-      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false);
+      const applied = ApplyComposites([compositeA], [modelA, modelB, modelC], (val) => val, false, timeZone);
       props.data = applied[0];
       render(<Tooltip {...props} />);
       expect(screen.getByRole('table')).toMatchSnapshot();
@@ -273,7 +274,7 @@ describe('Test Tooltips', () => {
       modelA.thresholdLevel = 0;
       modelB.thresholdLevel = 1;
       modelC.thresholdLevel = 0;
-      const applied = ApplyComposites([compositeB], [modelA, modelB, modelC], (val) => val, false);
+      const applied = ApplyComposites([compositeB], [modelA, modelB, modelC], (val) => val, false, timeZone);
       props.data = applied[0];
       // the displayMode comes from the applied composite, in this case there is one triggered metric
       if (applied[0].displayMode) {