+++ src/components/types.ts
@@ -28,6 +28,11 @@ export interface PolystatOptions {
   globalRegexPattern: string;
   globalShape: PolygonShapes;
   globalShowValueEnabled: boolean;
+  globalShowTimestampEnabled: boolean;
+  globalShowTimestampFormat: string;
+  globalShowTimestampFontSize: number;
+  globalShowTimestampPosition: TimestampPositions;
+  globalShowTimestampYOffset: number;
   globalShowTooltipColumnHeadersEnabled: boolean;
   globalTextFontAutoColor: string;
   globalTextFontAutoColorEnabled: boolean;
@@ -88,6 +93,8 @@ export interface PolystatModel {
   sanitizeURLEnabled: boolean;
   showName: boolean;
   showValue: boolean;
+  showTimestamp: boolean;
+  timestampFormatted: string;
   isComposite: boolean;
   members: PolystatModel[];
   triggerCache?: any; // holds animation frame info
@@ -184,4 +191,24 @@ export const FontFamilyOptionsLegacy = [
   { value: FontFamilies.HELVETICA_NEUE, label: 'Helvetica Neue' },
   { value: FontFamilies.ROBOTO, label: 'Roboto' },
   { value: FontFamilies.ROBOTO_MONO, label: 'Roboto Mono' },
+];
+
+export const ShowTimestampFormats = [
+  { value: 'HH:mm:ss', label: 'HH:mm:ss' },
+  { value: 'YYYY-MM-DD HH:mm', label: 'YYYY-MM-DD HH:mm' },
+  { value: 'YYYY-MM-DD HH:mm:ss', label: 'YYYY-MM-DD HH:mm:ss' },
+  { value: 'YYYY-MM-DD HH:mm:ss.SSS', label: 'YYYY-MM-DD HH:mm:ss.SSS' },
+  { value: 'MM/DD/YY h:mm:ss a', label: 'MM/DD/YY h:mm:ss a' },
+  { value: 'MMMM D, YYYY LT', label: 'MMMM D, YYYY LT' },
+  { value: 'YYYY-MM-DD', label: 'YYYY-MM-DD' },
+];
+
+export enum TimestampPositions {
+  BELOW_VALUE = 'below_value',
+  ABOVE_VALUE = 'above_value',
+}
+
+export const ShowTimestampPositions = [
+  { value: TimestampPositions.ABOVE_VALUE, label: 'Above Value' },
+  { value: TimestampPositions.BELOW_VALUE, label: 'Below Value' },
 ];