+++ src/data/composite_processor.ts
@@ -7,6 +7,7 @@ import { getTemplateSrv } from '@grafana/runtime';
 import { CompositeItemType, CompositeMetric } from '../components/composites/types';
 import { CUSTOM_SPLIT_DELIMITER } from './types';
 import { ApplyGlobalRegexPattern } from './processor';
+import { TimeFormatter } from './time_formatter';
 
 export const resolveCompositeTemplates = (
   metricComposites: CompositeItemType[],
@@ -109,6 +110,7 @@ const shallowClone = (item: PolystatModel): PolystatModel => {
     name: item.name,
     displayName: item.displayName,
     timestamp: item.timestamp,
+    timestampFormatted: item.timestampFormatted,
     prefix: item.prefix,
     suffix: item.suffix,
     color: item.color,
@@ -121,6 +123,7 @@ const shallowClone = (item: PolystatModel): PolystatModel => {
     customClickthroughTarget: '',
     showName: item.showName,
     showValue: item.showValue,
+    showTimestamp: item.showTimestamp,
     isComposite: item.isComposite,
     members: []
   };
@@ -132,6 +135,7 @@ export const ApplyComposites = (
   data: PolystatModel[],
   replaceVariables: InterpolateFunction,
   compositesGlobalAliasingEnabled: boolean,
+  timeZone: string,
   globalRegexPattern?: string,
 ): PolystatModel[] => {
   if (!composites) {
@@ -197,6 +201,11 @@ export const ApplyComposites = (
             seriesItem.customClickthroughTarget = aComposite.clickThroughCustomTarget;
             seriesItem.customClickthroughTargetEnabled = aComposite.clickThroughCustomTargetEnabled;
           }
+          // process the timestamp display
+          if (aComposite.showTimestampEnabled) {
+            seriesItem.timestampFormatted = TimeFormatter(timeZone, data[index].timestamp, aComposite.showTimestampFormat);
+            seriesItem.showTimestamp = true;
+          }
         }
       }
     }
@@ -234,6 +243,7 @@ export const ApplyComposites = (
       // add the composite setting for showing the name/value to the new cloned model
       clone.showName = aComposite.showName;
       clone.showValue = aComposite.showValue;
+      clone.showTimestamp = aComposite.showTimestampEnabled;
       clone.displayMode = aComposite.displayMode;
       clone.newTabEnabled = aComposite.clickThroughOpenNewTab;
       clone.sanitizeURLEnabled = aComposite.clickThroughSanitize;