+++ src/data/override_processor.test.ts
@@ -6,7 +6,7 @@ import { renderHook } from '@testing-library/react-hooks';
 
 import { ApplyOverrides, MatchOverride } from './override_processor';
 import { OverrideItemType } from 'components/overrides/types';
-import { PolystatModel } from '../components/types';
+import { PolystatModel, ShowTimestampFormats } from '../components/types';
 import { FieldConfigSource, FieldType, toDataFrame } from '@grafana/data';
 import { DataFrameToPolystat } from './processor';
 import { useTheme, useTheme2 } from '@grafana/ui';
@@ -31,7 +31,10 @@ describe('Test Overrides', () => {
     clickThroughOpenNewTab: false,
     clickThroughCustomTargetEnabled: false,
     clickThroughCustomTarget: '',
-    order: 0
+    order: 0,
+    showTimestampEnabled: false,
+    showTimestampFormat: ShowTimestampFormats[0].value,
+    showTimestampYOffset: 0,
   };
   const replaceVariables = (str: string) => {
     return str.replace(/\${noVariable}/g, 'noValue');
@@ -61,7 +64,7 @@ describe('Test Overrides', () => {
         },
         overrides: [],
       };
-      const { result } = renderHook(() => ApplyOverrides([overrideA], [modelA], fieldConfig, 'white', [], replaceVariables, useTheme(), useTheme2()));
+      const { result } = renderHook(() => ApplyOverrides([overrideA], [modelA], fieldConfig, 'white', [], replaceVariables, 'utc', useTheme(), useTheme2()));
       expect(result.all.length).toBe(1);
       const x = result.all[0] as PolystatModel[];
       const modified = x[0] as PolystatModel;