+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Entries
 
+## v2.1.9
+
+- NEW: Option to display timestamp of metric inside polygon, available in Global Configuration Section
+  - Show Timestamp Toggle on/off
+  - Preset time formats, and allows entry of arbitrary format
+  - Position of timestamp can be above or below the metric value (default is below)
+  - Y Offset can be specified to fine tune placement of timestamp displayed
+- Fixed alignment bug when name or value is disabled, will now align correctly (and use more space)
+
 ## v2.1.8
 
 - Tooltips will now display newest time of data received, and current time if none is supplied